backend/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==3.7.1
attrs==25.1.0
beautifulsoup4==4.12.2
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
fastapi==0.116.0
frozenlist==1.5.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.1.1
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
lxml==5.4.0
multidict==6.1.0
packaging==25.0
pluggy==1.6.0
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.2
pyparsing==3.2.3
pytest==8.4.1
python-docx==1.1.2
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9.1
schedule==1.2.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.23
starlette==0.46.2
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.2.0
urllib3==2.3.0
uvicorn==0.24.0
yarl==1.18.3
